+++ MinecraftServerLogsParsing/MinecraftServerLogsMapReduceJavaScript/src/main/java/org/timattt/jobs/onlyThread/OnlyThreadReducer.java
@@ -8,19 +8,19 @@ import org.apache.hadoop.mapreduce.Reducer;
 import java.io.IOException;
 import java.math.BigDecimal;
 
-public class OnlyThreadReducer extends Reducer<Text, DoubleWritable, Text, DoubleWritable> {
+public class OnlyThreadReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
     private final Text word = new Text();
-    private final DoubleWritable doubleWritable = new DoubleWritable();
+    private final IntWritable doubleWritable = new IntWritable();
 
-    public void reduce(Text key, Iterable<DoubleWritable> values, Context context) throws IOException, InterruptedException {
+    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
         BigDecimal result = new BigDecimal(0);
         int count = 0;
-        for (DoubleWritable day : values) {
+        for (IntWritable day : values) {
             result = result.add(BigDecimal.valueOf(day.get()));
             count++;
         }
         result = result.divide(BigDecimal.valueOf(count));
-        doubleWritable.set(result.doubleValue());
+        doubleWritable.set((int) result.doubleValue());
         context.write(key, doubleWritable);
     }
 }
